Move ScopeNamespaceLookup to its own file
Extract `ScopeNamespaceLookup` from `Scope.ts` into
`ScopeNamespaceLookup.ts`. Make `Program.getNamespaceContributors`
and `Program.getAggregateNamespaceContainer` public so the lookup
can call them across files without bracket-notation workarounds.
